Edited function in databaseHandler
